Alteration for AStart and DK
