+++ AStar.py
@@ -1,0 +1,10 @@
+'''
+Need Node to have an g and h value. IF value of found g' is less than g or h' less than h replace.
+f = g + h
+Check all nodes and choose node with minimum
+Node == cell space
+'''
+
+class Astart(Node):
+    def __init__(self):
+        pass
