+++ PathFinding.py
@@ -1,90 +1,36 @@
-
-from time import sleep
-from constants import INFITY, ROW_COMB, COL_COMB, ROW, COL
-'''
-Node class to hold each cells data
-'''
-class Node():
-	def __init__(self,posx: int,posy: int, value: int):
-		self.x = posx
-		self.y = posy
-		self.dist = value
-		self.visited = False
-		self.previousNode = None
-		self.type = None
-		self.inShort = False	
-
-	# Reset the node attributes
-	def reset(self):
-		self.dist = INFITY
-		self.visited = False
-		self.previousNode = None
-		self.type = None
-		self.inShort = None
-	
-	# Reset for rerun
-	def resetReRun(self):
-		self.dist = INFITY
-		self.visited = False
-		self.previousNode = None
-		self.inShort = None
-	
-	# Setters
-	def setType(self, newType):
-		self.type = newType
-	def setVisited(self):
-		self.visited = True
-	def setDist(self, val: int):
-		self.dist = val	
-	def setPrev(self, prev):
-		self.previousNode = prev
-	def setInShort(self):
-		self.inShort = True
-	
-	# Getters
-	def getType(self):
-		return self.type
-	def getVisited(self) -> bool:
-		return self.visited
-	def getDist(self)-> int:
-		return self.dist
-	def getPrev(self):
-		return self.previousNode
-	def getInShort(self):
-		return self.inShort
-	def getPosition(self) -> str:
-		return "(" + str(self.x) + "," + str(self.y) + ")"
-	
-	# Behaviors
-	def findNeighbours(self, graph: list) -> list:
-		neighbours = []
-		for i in range(len(ROW_COMB)):
-			x_value = self.x + ROW_COMB[i]
-			y_value = self.y + COL_COMB[i]
-			if (x_value < 0 or x_value >= ROW) or (y_value < 0 or y_value >= COL):
-				pass
-			else:
-				possible_neigh = graph[x_value][y_value]
-				if possible_neigh.getType() != "Wall" and possible_neigh.getVisited() != True:
-					neighbours.append(possible_neigh)
-		return neighbours
+from constants import INFITY
+from DkAlgo import DKNode
+from utils import Node
 
 '''Graph class for all grid related things'''
 class Graph():
 	def __init__(self, row, cols):
-		self._graph = self._initializeNodes(row, cols)
+		# NOTE: By default initialize as DK
+		self._graph = self._initializeDK(row, cols)
 		self.source = None
 		self.destination = None
 
 	# Initialize as nodes
-	def _initializeNodes(self, row, col):
+	def _initializeDK(self, row, col):
 		grid = []
 		for i in range(row):
 			temp = []
 			for j in range(col):
-				temp.append(Node(i, j, INFITY))
+				temp.append(DKNode(i, j, INFITY))
 			grid.append(temp)
 		return grid
+
+	# FIXME: Make init for A*
+	'''
+	def _initializeDK(self, row, col):
+		grid = []
+		for i in range(row):
+			temp = []
+			for j in range(col):
+				temp.append(DKNode(i, j, INFITY))
+			grid.append(temp)
+		return grid
+	'''
 	
 	# Setters
 	def setSource(self, row: int, col: int):
